+++ src/main/java/com/brick/demo/security/JwtRequestFilter.java
@@ -1,5 +1,6 @@
-package com.brick.demo.auth.jwt;
+package com.brick.demo.security;
 
+import com.brick.demo.auth.jwt.TokenProvider;
 import com.brick.demo.common.CustomException;
 import com.brick.demo.common.ErrorDetails;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,7 +17,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.util.AntPathMatcher;
 import org.springframework.util.StringUtils;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 @RequiredArgsConstructor
